% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{f_operator}
\usage{
f_operator(f)
}
\arguments{
\item{a}{factor 1}
}
\value{
A function
}
\description{
Catches error conditions for input functions
}
\examples{
sqrt.3 <- f_operator(sqrt.2)
\dontrun{
xxxxx
}
}
