% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_2.R
\name{log_2}
\alias{log_2}
\title{log_2}
\usage{
log_2(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
a number or, if x < 0, message = "Error: negative input, NA introduced!"
}
\description{
A robust log function
}
\examples{
log_2(-1)

}
