% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square2.R
\name{sqrt.2}
\alias{sqrt.2}
\title{sqrt.2}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
a number or, if x < 0, message = "Error: negative input, NA introduced!"
}
\description{
A robust square root function
}
\examples{
sqrt.2(-1)

}
