% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt2.R
\name{sqrt2}
\alias{sqrt2}
\title{sqrt2}
\usage{
sqrt2(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
a number or, if x < 0, message = "Error: negative input, NA introduced!"
}
\description{
A robust square root function
}
\examples{
sqrt2(-1)

}
