% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_2.R
\name{sqrt_2}
\alias{sqrt_2}
\title{sqrt_2}
\usage{
sqrt_2(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
a number or, if x < 0, message = "Error: negative input, NA introduced!"
}
\description{
A robust square root function
}
\examples{
sqrt_2(-1)

}
