% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{f_operator}
\usage{
f_operator(f)
}
\arguments{
\item{f}{A function sqrt_2 or log_2}
}
\value{
a function that can catch error condition and keep going
}
\description{
catches error conditions for input functions
}
\examples{
sqrt_3 = f_operator(sqrt_2)
set.seed(2)
(y = lapply(runif(2,-10,10), sqrt_3))

}
